# server/swagger.yaml
openapi: 3.0.0
info:
  title: OneUp API
  description: API contract for the OneUp project.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /rewards/{id}:
    get:
      summary: Get specific reward by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Reward details
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  value:
                    type: integer
                    example: 100
                  user_id:
                    type: integer
                    example: 1
                  created_at:
                    type: string
                    format: date-time
                    example: "2025-06-12T12:00:00Z"
                  updated_at:
                    type: string
                    format: date-time
                    example: "2025-06-12T12:00:00Z"
    put:
      summary: Update specific reward by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 123
                value:
                  type: integer
                  example: 200
      responses:
        '200':
          description: Reward updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 123
                  value:
                    type: integer
                    example: 200
                  user_id:
                    type: integer
                    example: 1
                  created_at:
                    type: string
                    format: date-time
                    example: "2025-06-12T12:00:00Z"
                  updated_at:
                    type: string
                    format: date-time
                    example: "2025-06-12T12:00:00Z"
    delete:
      summary: Delete specific reward by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 123
      responses:
        '204':
          description: Reward deleted successfully
...